requirements.txt
flask==2.2.5
flask[async]==2.2.5
nltk
numpy
pandas
scikit-learn
apify-client
Werkzeug==2.3.7
jinja2==3.1.2
flask-login==0.6.2
flask_migrate==4.0.4
WTForms==3.0.1
flask_wtf==1.2.1
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.21
email_validator==2.0.0
flask-restx==1.1.0
Pillow>=9.0.0
plotly
bokeh
python-dotenv>=0.21.0,<1.0
requests
gunicorn==20.1.0
Flask-Minify==0.37
six
weasyprint==57.0
pydyf==0.5.0
pydantic>=1.10
transformers
torch
sentencepiece
protobuf>=3.20.3,<6.0.0
# googletrans==4.0.0rc1
flask_mysqldb
psycopg2-binary
langchain>=0.3.26
langchain-community==0.3.27
chromadb==1.0.15
sentence-transformers==4.1.0
langgraph==0.5.3
langsmith==0.4.7
langchain-anthropic==0.3.17
langgraph-supervisor==0.0.27
langchain-tavily==0.2.10
tf-keras==2.18.0
apscheduler==3.11.0
pypdf
langchain-huggingface
langchain-chroma